+++ src/components/Page/components/Header/Header.tsx
@@ -52,9 +52,13 @@ export default class Header extends React.PureComponent<Props, State> {
     const breadcrumbMarkup =
       breadcrumbs.length > 0 ? <Breadcrumbs breadcrumbs={breadcrumbs} /> : null;
 
+    const primary =
+      primaryAction &&
+      (primaryAction.primary === undefined ? true : primaryAction.primary);
+
     const primaryActionMarkup = primaryAction ? (
       <div className={styles.PrimaryAction}>
-        {buttonsFrom(primaryAction, {primary: true})}
+        {buttonsFrom(primaryAction, {primary})}
       </div>
     ) : null;